+++ RPG/src/Crate.java
@@ -1,7 +1,7 @@
 
 public class Crate extends GameObject {
 
-	public Crate(double x, double y, double w, double h) {
+	public Crate(double x, double y, int w, int h) {
 		super(x, y, w, h, false, false, 25, "sprites/crate.png");
 
 	}
