+++ RPG/src/ExplosiveBarrel.java
@@ -2,7 +2,7 @@
 public class ExplosiveBarrel extends GameObject {
 
 
-	public ExplosiveBarrel(double x, double y, double w, double h) {
+	public ExplosiveBarrel(double x, double y, int w, int h) {
 		super(x, y, w, h, false, false, 25, "sprites/flask_red.png");
 	}
